+++ frontend/app.vue
@@ -1,3 +1,10 @@
+<script setup lang="ts">
+import { useAuthStore } from './stores/AuthStore';
+
+const authStore = useAuthStore();
+authStore.initializeStore();
+</script>
+
 <template>
   <TheNavbar />
   <NuxtPage />