+++ frontend/nuxt.config.ts
@@ -5,6 +5,6 @@ export default defineNuxtConfig({
       apiBase: 'http://127.0.0.1:8000',
     },
   },
-  modules: ["@vueuse/nuxt", "nuxt-icon"],
+  modules: ["@vueuse/nuxt", "nuxt-icon", '@pinia/nuxt',],
   ssr: false,
 })